refactor(semver): add post targets
